backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["kakakoi"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "==0.17.6"}
strawberry-graphql = "^0.114.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.19.0"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
bandit = "^1.7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 127
exclude = '''
(
    migrations
    | .mypy_cache
    | .pytest_cache
    | .venv
)
'''

[tool.isort]
line_length = 127
profile = "black"
